+++ driver-core/src/main/java/com/datastax/driver/core/DataType.java
@@ -58,7 +58,8 @@ public abstract class DataType {
         LIST      (32, List.class),
         SET       (34, Set.class),
         MAP       (33, Map.class),
-        UDT       (40, UDTValue.class), // TODO: we currently have no protocol id, so using a fake one to not throw off the logic below.
+        UDT       (48, UDTValue.class),
+        TUPLE     (49, TupleValue.class),
         CUSTOM    (0,  ByteBuffer.class);
 
         final int protocolId;
@@ -131,6 +132,7 @@ public abstract class DataType {
          *   <tr><td>TIMESTAMP     </td><td>Date</td></tr>
          *   <tr><td>UUID          </td><td>UUID</td></tr>
          *   <tr><td>UDT           </td><td>UDTValue</td></tr>
+         *   <tr><td>TUPLE         </td><td>TupleValue</td></tr>
          *   <tr><td>VARCHAR       </td><td>String</td></tr>
          *   <tr><td>VARINT        </td><td>BigInteger</td></tr>
          *   <tr><td>TIMEUUID      </td><td>UUID</td></tr>
@@ -153,7 +155,7 @@ public abstract class DataType {
     private static final Map<Name, DataType> primitiveTypeMap = new EnumMap<Name, DataType>(Name.class);
     static {
         for (Name name : Name.values()) {
-            if (!name.isCollection() && name != Name.CUSTOM && name != Name.UDT)
+            if (!name.isCollection() && name != Name.CUSTOM && name != Name.UDT && name != Name.TUPLE)
                 primitiveTypeMap.put(name, new DataType.Native(name));
         }
     }
@@ -168,7 +170,7 @@ public abstract class DataType {
         switch (name) {
             case CUSTOM:
                 String className = CBUtil.readString(buffer);
-                return CassandraTypeParser.isUserType(className)
+                return CassandraTypeParser.isUserType(className) || CassandraTypeParser.isTupleType(className)
                      ? CassandraTypeParser.parseOne(className)
                      : custom(className);
             case LIST:
@@ -385,16 +387,6 @@ public abstract class DataType {
     }
 
     /**
-     * Returns a User Defined Type (UDT).
-     *
-     * @param definition the description of the type's keyspace, name and fields.
-     * @return the UDT defined by {@code definition}.
-     */
-    public static DataType userType(UDTDefinition definition) {
-        return new UserType(definition);
-    }
-
-    /**
      * Returns the name of that type.
      *
      * @return the name of that type.
@@ -422,16 +414,6 @@ public abstract class DataType {
      */
     public List<DataType> getTypeArguments() {
         return Collections.<DataType>emptyList();
-    }
-
-    /**
-     * Returns their definition for user defined types (UDT).
-     *
-     * @return the definition of a user defined type or {@code null} for any other
-     * type.
-     */
-    public UDTDefinition getUDTDefinition() {
-        return null;
     }
 
     /**
@@ -678,46 +660,6 @@ public abstract class DataType {
                 return String.format("%s<%s, %s>", name, typeArguments.get(0), typeArguments.get(1));
             else
                 return String.format("%s<%s>", name, typeArguments.get(0));
-        }
-    }
-
-    private static class UserType extends DataType {
-
-        private final UDTDefinition definition;
-
-        private UserType(UDTDefinition definition) {
-            super(DataType.Name.UDT);
-            this.definition = definition;
-        }
-
-        @SuppressWarnings("unchecked")
-        @Override
-        TypeCodec<Object> codec(int protocolVersion) {
-            return (TypeCodec)TypeCodec.udtOf(definition);
-        }
-
-        @Override
-        public UDTDefinition getUDTDefinition() {
-            return definition;
-        }
-
-        @Override
-        public final int hashCode() {
-            return Arrays.hashCode(new Object[]{ name, definition });
-        }
-
-        @Override
-        public final boolean equals(Object o) {
-            if(!(o instanceof DataType.UserType))
-                return false;
-
-            DataType.UserType d = (DataType.UserType)o;
-            return name == d.name && definition.equals(d.definition);
-        }
-
-        @Override
-        public String toString() {
-            return Metadata.escapeId(definition.getKeyspace()) + '.' + Metadata.escapeId(definition.getName());
         }
     }