+++ driver-core/src/test/java/com/datastax/driver/core/StateListenerTest.java
@@ -25,7 +25,7 @@ import static com.datastax.driver.core.TestUtils.*;
  */
 public class StateListenerTest {
 
-    @Test(groups = "long")
+    @Test(groups = "short")
     public void listenerTest() throws Throwable {
 
         CCMBridge.CCMCluster c = CCMBridge.buildCluster(1, Cluster.builder());
@@ -79,6 +79,9 @@ public class StateListenerTest {
 
         public void onUp(Host host) {
             ups++;
+        }
+
+        public void onSuspected(Host host) {
         }
 
         public void onDown(Host host) {